+++ apps/web-v2/src/features/agents/components/create-edit-agent-dialogs/create-agent-dialog.tsx
@@ -30,7 +30,8 @@ import { DeepAgentConfiguration } from "@/types/deep-agent";
 import {
   DEFAULT_FORM_CONFIG,
   prepareConfigForSaving,
-  isValidDeepAgentConfiguration,
+  handleCopyConfig,
+  handlePasteConfig,
 } from "./utils";
 
 interface CreateAgentDialogProps {
@@ -222,83 +223,6 @@ export function CreateAgentDialog({
     null,
   );
 
-  const handleCopyConfig = async () => {
-    if (!formRef) return;
-
-    const formData = formRef.getValues();
-    const configToCopy = {
-      name: formData.name,
-      metadata: {
-        description: formData.description,
-      },
-      config: {
-        configurable: formData.config,
-      },
-    };
-
-    try {
-      await navigator.clipboard.writeText(
-        JSON.stringify(configToCopy, null, 2),
-      );
-      toast.success("Agent configuration copied to clipboard", {
-        richColors: true,
-      });
-    } catch {
-      toast.error("Failed to copy configuration", {
-        richColors: true,
-      });
-    }
-  };
-
-  const handlePasteConfig = async () => {
-    if (!formRef) return;
-
-    try {
-      const clipboardText = await navigator.clipboard.readText();
-      const parsedConfig = JSON.parse(clipboardText);
-
-      // Validate the structure
-      if (typeof parsedConfig !== "object" || parsedConfig === null) {
-        throw new Error("Invalid configuration format");
-      }
-
-      // Expect only the new agent structure format
-      if (!(parsedConfig.metadata && parsedConfig.config?.configurable)) {
-        throw new Error("Invalid configuration format");
-      }
-
-      const name = parsedConfig.name as string | undefined;
-      const description = parsedConfig.metadata.description as
-        | string
-        | undefined;
-      const config = parsedConfig.config.configurable as
-        | Record<string, unknown>
-        | undefined;
-
-      // Set form values
-      if (name) {
-        formRef.setValue("name", name);
-      }
-      if (description) {
-        formRef.setValue("description", description);
-      }
-      if (isValidDeepAgentConfiguration(config)) {
-        formRef.setValue("config", config);
-      }
-
-      toast.success("Agent configuration pasted successfully", {
-        richColors: true,
-      });
-    } catch (_error) {
-      toast.error(
-        "Failed to paste configuration. Please ensure it's valid JSON format.",
-        {
-          richColors: true,
-        },
-      );
-    }
-  };
-
   const lastOpen = useRef(open);
   useLayoutEffect(() => {
     if (lastOpen.current !== open && open) {
@@ -327,7 +251,7 @@ export function CreateAgentDialog({
               <Button
                 variant="outline"
                 size="sm"
-                onClick={handleCopyConfig}
+                onClick={() => handleCopyConfig(formRef)}
                 className="flex items-center gap-2"
                 disabled={!formRef}
               >
@@ -337,7 +261,7 @@ export function CreateAgentDialog({
               <Button
                 variant="outline"
                 size="sm"
-                onClick={handlePasteConfig}
+                onClick={() => handlePasteConfig(formRef)}
                 className="flex items-center gap-2"
                 disabled={!formRef}
               >
